+++ __test__/test-cli-unit.test.js
@@ -1,9 +1,8 @@
-const path = require('path');
-
-const exec = require('shelljs').exec,
+const path = require('path'),
   sdk = require('postman-collection'),
-  fs = require('fs'), 
-  { resetAndReimportUrlUpdator, CustomError } = require('./utils/helperMethods.js');
+  fs = require('fs'),
+  { resetAndReimportUrlUpdator, CustomError } = require('./utils/helperMethods.js'),
+  test = require('../src/index.js')
 
 
 let postman_url_updater,
@@ -12,7 +11,7 @@ let postman_url_updater,
   mockConsoleLog = jest.spyOn(console, 'log').mockImplementation(() => { });
 
 describe('Validate index.js unit tests', () => {
-  describe('Validate cli output messages', () => {
+  describe('Validate cli output messages for non interactive', () => {
 
     afterEach(() => {
       jest.clearAllMocks();
@@ -23,8 +22,17 @@ describe('Validate index.js unit tests', () => {
     })
 
     it('Should show help if arguments are missing', async () => {
-      resetAndReimportUrlUpdator({ c: undefined, r: undefined, w: undefined, s: undefined, p: undefined });
-      expect(() => { require('../index.js') }).toThrowErrorMatchingSnapshot()
+      resetAndReimportUrlUpdator({ c: undefined, r: undefined, w: undefined, s: undefined, i: undefined });
+
+      expect(() => { require('../src/index.js') }).not.toThrowError()
+      expect(mockProcessExit).not.toBeCalled();
+      expect(mockConsoleError).not.toBeCalled();
+      expect(mockProcessExit).toMatchSnapshot();
+      expect(mockConsoleError).toMatchSnapshot();
+      expect(mockConsoleLog).not.toBeCalled();
+      expect(mockConsoleLog).toMatchSnapshot();
+
+      await expect(async () => { await require('../src/index.js').startConvert() }).rejects.toThrowErrorMatchingSnapshot()
       expect(mockProcessExit).toBeCalled();
       expect(mockConsoleError).toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -34,8 +42,18 @@ describe('Validate index.js unit tests', () => {
     });
 
     it('Should show help if -c argument is missing', async () => {
-      resetAndReimportUrlUpdator({ c: undefined, r: "{{baseURL}}/{{path}}", w: "new_collection.json", s: undefined, p: undefined });
-      expect(() => { require('../index.js') }).toThrowErrorMatchingSnapshot()
+      resetAndReimportUrlUpdator({ c: undefined, r: "{{baseURL}}/{{path}}", w: "new_collection.json", s: undefined, i: undefined });
+
+      expect(() => { require('../src/index.js') }).not.toThrowError()
+      expect(mockProcessExit).not.toBeCalled();
+      expect(mockConsoleError).not.toBeCalled();
+      expect(mockProcessExit).toMatchSnapshot();
+      expect(mockConsoleError).toMatchSnapshot();
+      expect(mockConsoleLog).not.toBeCalled();
+      expect(mockConsoleLog).toMatchSnapshot();
+
+      await expect(async () => { await require('../src/index.js').startConvert() }).rejects.toThrowErrorMatchingSnapshot()
+
       expect(mockProcessExit).toBeCalled();
       expect(mockConsoleError).toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -45,8 +63,18 @@ describe('Validate index.js unit tests', () => {
     });
 
     it('Should show help if -r argument is missing', async () => {
-      resetAndReimportUrlUpdator({ c: "test/collection/collection.json", r: undefined, w: "new_collection.json", s: undefined, p: undefined });
-      expect(() => { require('../index.js') }).toThrowErrorMatchingSnapshot()
+      resetAndReimportUrlUpdator({ c: "test/collection/collection.json", r: undefined, w: "new_collection.json", s: undefined, i: undefined });
+
+      expect(() => { require('../src/index.js') }).not.toThrowError()
+      expect(mockProcessExit).not.toBeCalled();
+      expect(mockConsoleError).not.toBeCalled();
+      expect(mockProcessExit).toMatchSnapshot();
+      expect(mockConsoleError).toMatchSnapshot();
+      expect(mockConsoleLog).not.toBeCalled();
+      expect(mockConsoleLog).toMatchSnapshot();
+
+      await expect(async () => { await require('../src/index.js').startConvert() }).rejects.toThrowErrorMatchingSnapshot()
+
       expect(mockProcessExit).toBeCalled();
       expect(mockConsoleError).toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -56,8 +84,18 @@ describe('Validate index.js unit tests', () => {
     });
 
     it('Should show help if -w argument is missing', async () => {
-      resetAndReimportUrlUpdator({ c: "test/collection/collection.json", r: "{{baseURL}}/{{path}}", w: undefined, s: "new_collection.json", p: undefined });
-      expect(() => { require('../index.js') }).toThrowErrorMatchingSnapshot()
+      resetAndReimportUrlUpdator({ c: "test/collection/collection.json", r: "{{baseURL}}/{{path}}", w: undefined, s: "new_collection.json", i: undefined });
+
+      expect(() => { require('../src/index.js') }).not.toThrowError()
+      expect(mockProcessExit).not.toBeCalled();
+      expect(mockConsoleError).not.toBeCalled();
+      expect(mockProcessExit).toMatchSnapshot();
+      expect(mockConsoleError).toMatchSnapshot();
+      expect(mockConsoleLog).not.toBeCalled();
+      expect(mockConsoleLog).toMatchSnapshot();
+
+      await expect(async () => { await require('../src/index.js').startConvert() }).rejects.toThrowErrorMatchingSnapshot()
+
       expect(mockProcessExit).toBeCalled();
       expect(mockConsoleError).toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -67,8 +105,8 @@ describe('Validate index.js unit tests', () => {
     });
 
     it('Should show file not found error if collection doesnt exists', async () => {
-      resetAndReimportUrlUpdator({ c: "test/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "new_collection.json", p: undefined });
-      postman_url_updater = require('../index.js')
+      resetAndReimportUrlUpdator({ c: "test/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "new_collection.json", i: undefined });
+      postman_url_updater = require('../src/index.js')
       expect(mockProcessExit).not.toBeCalled();
       expect(mockConsoleError).not.toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -76,7 +114,7 @@ describe('Validate index.js unit tests', () => {
       expect(mockConsoleLog).not.toBeCalled();
       expect(mockConsoleLog).toMatchSnapshot();
       jest.clearAllMocks()
-      expect(postman_url_updater.startConvert).toThrowErrorMatchingSnapshot();
+      await expect(postman_url_updater.startConvert).rejects.toThrowErrorMatchingSnapshot();
       expect(mockProcessExit).toBeCalled();
       expect(mockConsoleError).toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -86,8 +124,8 @@ describe('Validate index.js unit tests', () => {
     });
 
     it('Should show file saved message correctly with default path', async () => {
-      resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: undefined, p: undefined });
-      postman_url_updater = require('../index.js')
+      resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: undefined, i: undefined });
+      postman_url_updater = require('../src/index.js')
       expect(mockProcessExit).not.toBeCalled();
       expect(mockConsoleError).not.toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -95,7 +133,7 @@ describe('Validate index.js unit tests', () => {
       expect(mockConsoleLog).not.toBeCalled();
       expect(mockConsoleLog).toMatchSnapshot();
       jest.clearAllMocks()
-      postman_url_updater.startConvert();
+      await postman_url_updater.startConvert();
       expect(mockProcessExit).not.toBeCalled();
       expect(mockConsoleError).not.toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -104,8 +142,8 @@ describe('Validate index.js unit tests', () => {
     });
 
     it('Should show file saved message correctly with custom path', async () => {
-      resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "__test__/collection/output/new_collection.json", p: undefined });
-      postman_url_updater = require('../index.js')
+      resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "__test__/collection/output/new_collection.json", i: undefined });
+      postman_url_updater = require('../src/index.js')
       expect(mockProcessExit).not.toBeCalled();
       expect(mockConsoleError).not.toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -113,7 +151,7 @@ describe('Validate index.js unit tests', () => {
       expect(mockConsoleLog).not.toBeCalled();
       expect(mockConsoleLog).toMatchSnapshot();
       jest.clearAllMocks()
-      postman_url_updater.startConvert()
+      await postman_url_updater.startConvert()
       expect(mockProcessExit).not.toBeCalled();
       expect(mockConsoleError).not.toBeCalled();
       expect(mockProcessExit).toMatchSnapshot();
@@ -122,12 +160,36 @@ describe('Validate index.js unit tests', () => {
     });
 
     it('Should throw error if exception thrown doesnt contain error code ENOENT', async () => {
-      resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "new_collection.json", p: undefined });
-      postman_url_updater = require('../index.js')
+      resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "new_collection.json", i: undefined });
+      postman_url_updater = require('../src/index.js')
       jest.clearAllMocks()
       mockConsoleLog.mockImplementation(() => { throw new CustomError(256) });
-      expect(postman_url_updater.startConvert).toThrowErrorMatchingSnapshot()
+      await expect(postman_url_updater.startConvert).rejects.toThrowErrorMatchingSnapshot()
     });
+  })
+
+  describe('Validate cli output messages for interactive', () => {
+
+    afterEach(() => {
+      jest.clearAllMocks();
+    });
+
+    afterAll(() => {
+      jest.restoreAllMocks();
+    })
+
+    it('Should call inquirer if i not provided', async () => {
+      let inquirerSpy = resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "__test__/collection/output/new_collection.json", i: true });
+      await require('../src/index.js').startConvert();
+      expect(inquirerSpy).toHaveBeenCalled()
+    });
+
+    it('Should not call inquirer if i not provided', async () => {
+      let inquirerSpy = resetAndReimportUrlUpdator({ c: "__test__/collection/collection.json", r: "{{baseURL}}/{{path}}", w: "{{baseURL}}/{{path}}", s: "__test__/collection/output/new_collection.json", i: false });
+      await require('../src/index.js').startConvert();
+      expect(inquirerSpy).not.toHaveBeenCalled()
+    });
+    
   })
 
   describe('Validate collection changes', () => {
@@ -136,8 +198,8 @@ describe('Validate index.js unit tests', () => {
 
     it('validate no changes happens when there is no match', async () => {
       const sourceCollection = new sdk.Collection(JSON.parse(fs.readFileSync(collectionPath).toString())).toJSON();
-      resetAndReimportUrlUpdator({ c: collectionPath, r: "https://localhost:23456/api/v1/{{path}}", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, p: undefined });
-      require('../index.js').startConvert();
+      resetAndReimportUrlUpdator({ c: collectionPath, r: "https://localhost:23456/api/v1/{{path}}", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, i: undefined });
+      await require('../src/index.js').startConvert();
       const outputCollection = new sdk.Collection(JSON.parse(fs.readFileSync(outputCollectionPath).toString())).toJSON();
       expect(outputCollection.item[0].item[0].request).toStrictEqual(sourceCollection.item[0].item[0].request);
       expect(outputCollection.item[1].item[0].item[0].item[0].item[0].item[0].item[0].item[0].request)
@@ -160,8 +222,8 @@ describe('Validate index.js unit tests', () => {
       }
 
       const sourceCollection = new sdk.Collection(JSON.parse(fs.readFileSync(collectionPath).toString())).toJSON();
-      resetAndReimportUrlUpdator({ c: collectionPath, r: "https://www.testdomain0.ie", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, p: undefined });
-      require('../index.js').startConvert();
+      resetAndReimportUrlUpdator({ c: collectionPath, r: "https://www.testdomain0.ie", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, i: undefined });
+      await require('../src/index.js').startConvert();
       const outputCollection = new sdk.Collection(JSON.parse(fs.readFileSync(outputCollectionPath).toString())).toJSON();
 
       updatedRequestObject(sourceCollection.item[0].item[0], 1, 1);
@@ -189,8 +251,8 @@ describe('Validate index.js unit tests', () => {
       }
 
       const sourceCollection = new sdk.Collection(JSON.parse(fs.readFileSync(collectionPath).toString())).toJSON();
-      resetAndReimportUrlUpdator({ c: collectionPath, r: "www.testdomain0.ie", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, p: undefined });
-      require('../index.js').startConvert();
+      resetAndReimportUrlUpdator({ c: collectionPath, r: "www.testdomain0.ie", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, i: undefined });
+      await require('../src/index.js').startConvert();
       const outputCollection = new sdk.Collection(JSON.parse(fs.readFileSync(outputCollectionPath).toString())).toJSON();
 
       updatedRequestObject(sourceCollection.item[0].item[0], 1, 1);
@@ -218,8 +280,8 @@ describe('Validate index.js unit tests', () => {
       }
 
       const sourceCollection = new sdk.Collection(JSON.parse(fs.readFileSync(collectionPath).toString())).toJSON();
-      resetAndReimportUrlUpdator({ c: collectionPath, r: "www.testdomain0.ie/testpath/:pathvariable1-1", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, p: undefined });
-      require('../index.js').startConvert();
+      resetAndReimportUrlUpdator({ c: collectionPath, r: "www.testdomain0.ie/testpath/:pathvariable1-1", w: "{{baseURL}}/{{path}}", s: outputCollectionPath, i: undefined });
+      await require('../src/index.js').startConvert();
       const outputCollection = new sdk.Collection(JSON.parse(fs.readFileSync(outputCollectionPath).toString())).toJSON();
 
       updatedRequestObject(sourceCollection.item[0].item[0], 1, 1);
@@ -250,8 +312,8 @@ describe('Validate index.js unit tests', () => {
       }
 
       const sourceCollection = new sdk.Collection(JSON.parse(fs.readFileSync(collectionPath).toString())).toJSON();
-      resetAndReimportUrlUpdator({ c: collectionPath, r: "/:pathvariable1-1/path/request1-1?query1-1-0", w: "/:pathvariable1-1/path/request1-1?query1-1-0=queryvalue1-1-0-0&test", s: outputCollectionPath, p: undefined });
-      require('../index.js').startConvert();
+      resetAndReimportUrlUpdator({ c: collectionPath, r: "/:pathvariable1-1/path/request1-1?query1-1-0", w: "/:pathvariable1-1/path/request1-1?query1-1-0=queryvalue1-1-0-0&test", s: outputCollectionPath, i: undefined });
+      await require('../src/index.js').startConvert();
       const outputCollection = new sdk.Collection(JSON.parse(fs.readFileSync(outputCollectionPath).toString())).toJSON();
 
       updatedRequestObject(sourceCollection.item[0].item[0], 1, 1);
@@ -282,9 +344,9 @@ describe('Validate index.js unit tests', () => {
       resetAndReimportUrlUpdator({
         c: collectionPath,
         r: "https://www.testdomain0.ie/testpath/:pathvariable1-1/path/request1-1?query1-1-0=queryvalue1-1-0&query1-1-1=queryvalue1-1-1",
-        w: "{{protocol}}://{{baseURL}}/v1/{{path}}", s: outputCollectionPath, p: undefined
+        w: "{{protocol}}://{{baseURL}}/v1/{{path}}", s: outputCollectionPath, i: undefined
       });
-      require('../index.js').startConvert();
+      await require('../src/index.js').startConvert();
 
       const outputCollection = new sdk.Collection(JSON.parse(fs.readFileSync(outputCollectionPath).toString())).toJSON();